+++ src/main/java/net/minecraft/client/renderer/debug/ChunkDebugRenderer.java
@@ -69,7 +69,7 @@ public class ChunkDebugRenderer implements DebugRenderer.SimpleDebugRenderer {
                 for (String part : parts) {
                     Gizmos.billboardText(
                             part,
-                            new Vec3(SectionPos.sectionToBlockCoord(pos.x(), 8), y + yOffset, SectionPos.sectionToBlockCoord(pos.z(), 8)),
+                            new Vec3(SectionPos.sectionToBlockCoord((int)pos.x(), 8), y + yOffset, SectionPos.sectionToBlockCoord((int)pos.z(), 8)),
                             TextGizmo.Style.whiteAndCentered().withScale(2.4F)
                         )
                         .setAlwaysOnTop();
