+++ src/main/java/net/minecraft/server/commands/ExecuteCommand.java
@@ -459,7 +459,7 @@ public class ExecuteCommand {
 
     private static boolean isChunkLoaded(final ServerLevel level, final BlockPos pos) {
         ChunkPos chunkPos = ChunkPos.containing(pos);
-        LevelChunk chunk = level.getChunkSource().getChunkNow(chunkPos.x(), chunkPos.z());
+        LevelChunk chunk = level.getChunkSource().getChunkNow((int)chunkPos.x(), (int)chunkPos.z());
         return chunk == null ? false : chunk.getFullStatus() == FullChunkStatus.ENTITY_TICKING && level.areEntitiesLoaded(chunkPos);
     }
