+++ src/main/java/net/minecraft/world/entity/monster/cubemob/Slime.java
@@ -90,7 +90,7 @@ public class Slime extends AbstractCubeMob implements Enemy {
             }
 
             ChunkPos chunkPos = ChunkPos.containing(pos);
-            boolean slimeChunk = WorldgenRandom.seedSlimeChunk(chunkPos.x(), chunkPos.z(), worldGenLevel.getSeed(), 987234911L).nextInt(10) == 0;
+            boolean slimeChunk = WorldgenRandom.seedSlimeChunk((int)chunkPos.x(), (int)chunkPos.z(), worldGenLevel.getSeed(), 987234911L).nextInt(10) == 0;
             if (random.nextInt(10) == 0 && slimeChunk && pos.getY() < 40) {
                 return checkMobSpawnRules(type, level, spawnReason, pos, random);
             }
