+++ src/main/java/net/minecraft/world/level/chunk/status/ChunkStatusTasks.java
@@ -182,7 +182,7 @@ public class ChunkStatusTasks {
         final WorldGenContext context, final ChunkStep step, final StaticCache2D<GenerationChunkHolder> chunks, final ChunkAccess chunk
     ) {
         ChunkPos pos = chunk.getPos();
-        GenerationChunkHolder holder = chunks.get(pos.x(), pos.z());
+        GenerationChunkHolder holder = chunks.get((int)pos.x(), (int)pos.z());
         return CompletableFuture.supplyAsync(() -> {
             ProtoChunk protoChunk = (ProtoChunk)chunk;
             ServerLevel level = context.level();
